+++ changelog.md
@@ -2,7 +2,14 @@
 Notable changes to the `cc.isr.vi` solution are documented in this file using the 
 [Keep a Changelog] style. The dates specified are in coordinated universal time (UTC).
 
-[8.1.9264]: https://www.github.com/atecoder/dn.vi.ivi
+[8.1.9265]: https://www.github.com/atecoder/dn.vi.ivi
+
+## [8.1.9265] - 2025-05-14
+- Device Tsp Script Session Base Method
+  - Add a helper methods to build and load a script.
+  - Delete Object: Fix the code to correctly report the existence of the object.
+  - Script info: 
+    - Remove initialize and dynamic properties. It must be set specifically.
 
 ## [8.1.9264] - 2025-05-13
 - Lexemes: